pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vibecheck"
version = "0.1.0"
description = "Understand the code your AI wrote"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "vibecheck contributors"},
]
keywords = ["cli", "code-review", "ai", "education", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "litellm>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
vibecheck = "vibecheck.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["vibecheck*"]
